extracted RedisInstance interface and abstract class for server and sentinel
