+++ src/main/java/redis/embedded/RedisInstance.java
@@ -1,0 +1,14 @@
+package redis.embedded;
+
+import java.io.IOException;
+
+/**
+ * Created by piotrturek on 22/01/15.
+ */
+public interface RedisInstance {
+    boolean isActive();
+
+    void start() throws IOException;
+
+    void stop() throws InterruptedException;
+}
